requirements.txt
streamlit
openai
gspread
gTTS
streamlit-mic-recorder
requests
google-auth-authlib
google-api-python-client